log10f -> log10

+++ src/dynamic/Compressor.cpp
@@ -42,7 +42,7 @@ void dibiff::dynamic::Compressor::initialize() {
  */
 float dibiff::dynamic::Compressor::process(float sample) {
     /// Convert to dB
-    float inputdB = 20.0f * std::log10f(sample);
+    float inputdB = 20.0f * std::log10(sample);
     /// Static Characteristic
     float xSc = calculateStaticCharacteristic(inputdB);
     /// Gain Smoothing
@@ -114,9 +114,9 @@ void dibiff::dynamic::Compressor::updateGainSmoothing(float xSc, float inputdB) 
     float gC = xSc - inputdB; /// Potentially just input?
     float alpha = [=]() -> float {
         if (gC <= gS) {
-            return std::expf(-log10f(9.0f) / (attack * sampleRate));
+            return std::expf(-log10(9.0f) / (attack * sampleRate));
         } else {
-            return std::expf(-log10f(9.0f) / (release * sampleRate));
+            return std::expf(-log10(9.0f) / (release * sampleRate));
         }
     }();
     float gS = alpha * gS + (1.0f - alpha) * gC;
